[rules] Let snipers pick out leadership models, but not individual models
Being able to target individual models is overly fiddly considering that
various unit-based rules apply to a whole unit but would logically
belong to a specific model--for example the Medic rule. Targeting the
model in the unit which is modeled as a medic would logically remove
that rule from the unit, but there are no game mechanics for this and I
don't want to add one, in order to preserve modeling flexibility and
fairness for units without special modeling (perhaps the medic's equipment
is stowed in a backpack or something).