Templatize units' weapons so they can be defined only once
Now weapons are defined centrally in each army list with default hit
values that can be overridden by appending "[5+]" (for example) to
individual usages of the weapon in a unit entry.

Doing this cleans up tons and tons of inconsistencies in style and
formatting when the weapons were duplicated across unit entries.

This is done for the two existing armies: Orc Raiders and Planetary
Defense Forces.