[rules] Make Defend action Life bonuses only apply against shooting attacks
This provides an additional incentive to initiate melee combat against a
deeply entrenched enemy in cover: surviving the charge will expose the
enemy unit without any of its defensive bonuses, making its models much
easier to kill. This should be especially deadly when paired with Assault
weapons, which can be fired at point-blank range during a Fight action.
This is exactly as it should be.