Use auto-generated URL pointing to latest versions of assets in README
This will save the hassle of manually updating the README file after
every release, as long as the filenames don't change.

Accordingly, this entails losing the version number in the filenames.
This is probably fine as users will generally not care and may even
prefer the filenames to not have version numbers in them.

Fixes #22