[rules] Add basic psychic combat mechanics for casting and nullifying
Now powers go off automatically without rolling, but nullifying requires
a Will roll, and resisting nullification requires a Will roll as well.

Additionally, the Unstable rule applied to psychic Will rolls is
increased to 2, so there is now a 20% chance of killing yourself for
resisting or pushing through nullification.

This makes psychics much more interesting. Because psychic power are now
automatically cast if not nullified, there is a strong incentive to take
at least one psychic model in your army just for counter-psychics.

And Psychic combar is now interesting: both the caster and the nullifier
expose themselves to risk for continuing the combat.

Related to #46