[rules] Will roll to give an Order now only needed to extend range
Previously all orders required a Will roll to succeed. Now, the order
automatically takes effect, but the Leadership unit issuing it can pass
a Will roll to extend the range to 18". However if they fail the roll,
the order is wasted, counting against the unit's maximum number of
orders.

Related to #46