[rules] Apply Bulky rule to PDF and Orc units that need it
This also applies to bikers and cavalry, which I find somewhat
hilarious. Imagine bikes or horses pouring out of a transport vehicle!
Their Bulky values are quite high so you won't be able to fit many in
there, but still, you can do it, if your really want to, I guess. :)