[rules] Generally Reduce Move values across the board
Some distances in the basic rules are also reduces to 5" or
multiples of 5".

The general idea here is to not have everything zooming around the
board quite so fast, which is something that heavily favors fast
armies and experienced players who can use speed to their advantage
and dominate the flow of the game.

If you want to get where you're going quickly, you should have to use
a transport vehicle or be a vehicle, essentially. This makes those
units more important, which is exactly as it should be in an era of
semi to fully mechanized warfare.

Now:
- Most slow wheeled/tracked vehicles have Move 8
- Medium-speed ones have Move 10
- Fast ones have Move 12
- Walkers have around Move 6-7
- Infantry generally have Move 5, with light troops and infiltrators
  often having Move 6
- Heavily-armored infantry generally have Move 4
- Very fast infantry have Move 6 or even 7

Fixes #49
Fixes #64