[rules] Each advanced shooting task requires a will roll now
Rather than the Will roll giving you the ability to to any and all of
them. The fancier you want your shooting to be, the more you have to
risk! Because each failed roll imposes an additional -1 Shoot penalty
for the remainder of the action due to the confusion of command and
targeting getting mixed up for a little while.