[rules] Add an advanced shooting task to slightly increase weapon ranges
This is designed for the case where the unit you want to shoot at is
just like an inch outisde the weapon's maximum range and you're
tearing your hair out over it! Now you can try to have the unit fire
past their weapons' effective ranges, but with a large Shoot penalty.
And of course failure will impose a further Shoot penalty on all the
unit's shooting for that action.

As part of this change, Vehicle/Monster units are now granted the
ability to perform all advanced shooting actions for free, since they
never make Will rolls.