[rules] Remove rule to limit wounded Vehicle/Monster movement
As written, this rule made no sense, and would have limited the movement
of an un-wounded Vehicle or Monster with a Move value higher than its
starting Tough value.

Let's remove it and migrate the basic idea later to a set of advanced
Vehicle/Monster rules that allow for reduced mobility, destroyed
weapons, etc.