[rules] Let PDF units take the Missile Launcher
Also remove the Auto Grenade Launcher, as one of the Missile Launcher's
fire modes is very similar.

Finaly, raise the price of the the Missile Launcher by 10, relative to
the old price of the Rocket Launcher weapon that it replaces. The old
Rocket Launcher was probably too cheap in the first place given its
anti-tank capabiliy, and the Missile Launcher that replaces it is more
versatile with its blast mode.

Fixes #82